Basic test harness
