+++ bin/server
@@ -1,0 +1,4 @@
+#!/usr/bin/env sh
+
+go build -o /go/bin/wordcount
+exec /go/bin/wordcount
