+++ endpoints.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+    "net/http"
+)
+
+func DefaultEndpoint(w http.ResponseWriter, r *http.Request) {
+    w.Write([]byte("Hello World!\n"))
+}
