+++ middleware.go
@@ -1,0 +1,13 @@
+package main
+
+import (
+    "log"
+    "net/http"
+)
+
+func LoggingMiddleware(next http.Handler) http.Handler {
+    return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+        log.Println(r.RequestURI)
+        next.ServeHTTP(w, r)
+    })
+}
